// src/types/preRegistration.ts
export interface PreRegistrationData {
    schoolId: number;
    schoolName: string;
    schoolAddress: string | null;
    schoolCity: string | null;
    schoolState: string | null;
    schoolPincode: string | null;
    schoolWebsite: string | null;
    firstName: string | null;
    surname: string | null;
    coordinatorMobileNo: string | null;
    contactEmail: string;
  }